requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
black==22.6.0
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
flake8==4.0.1
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
fonttools==4.33.3
idna==3.3
investpy==1.0.8
isort==5.10.1
kiwisolver==1.4.2
lxml==4.9.0
matplotlib==3.5.2
mccabe==0.6.1
mypy==0.960
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
ta==0.10.1
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.9
